Refactor: Restructure and Improve the Application
This commit introduces a major refactoring of the application to improve its structure, maintainability, and overall quality.

The key changes include:
- Restructured the project into a `src` directory with a more organized folder structure.
- Broke down the monolithic `index.tsx` into smaller, reusable React components.
- Improved type safety by introducing a `types.ts` file and replacing `any` types.
- Pinned all dependencies in `package.json` to specific versions.
- Fixed a bug in the alarm functionality to ensure alarms persist across page reloads.
