+++ index.html
@@ -8,7 +8,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;600;700;900&display=swap" rel="stylesheet">
     <title>Pequenos Passos - Para uma Vida Extraordinária</title>
-    <link rel="stylesheet" href="/index.css">
+    <link rel="stylesheet" href="/src/index.css">
   <script type="importmap">
 {
   "imports": {
@@ -22,6 +22,6 @@
 </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/index.tsx"></script>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
