+++ package.json
@@ -9,13 +9,13 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "react-dom": "^19.1.1",
-    "react": "^19.1.1",
-    "@google/genai": "latest"
+    "react-dom": "19.1.1",
+    "react": "19.1.1",
+    "@google/genai": "1.17.0"
   },
   "devDependencies": {
-    "@types/node": "^22.14.0",
-    "typescript": "~5.8.2",
-    "vite": "^6.2.0"
+    "@types/node": "22.14.0",
+    "typescript": "5.8.2",
+    "vite": "6.2.0"
   }
 }
